requirements.txt
google-generativeai==0.8.4
pydantic==2.11.1
tqdm==4.67.1
neo4j==5.28.1
beautifulsoup4==4.13.3
python-dotenv==1.1.0
argparse==1.4.0
openai==1.69.0